All BDD and unit testing passing

+++ src/main/java/br/com/nubank/capital/gain/delivery/dto/OperationDto.java
@@ -1,6 +1,6 @@
 package br.com.nubank.capital.gain.delivery.dto;
 
-import br.com.nubank.capital.gain.domain.enums.OperationType;
+import br.com.nubank.capital.gain.domain.entity.enums.OperationType;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.math.BigDecimal;
@@ -14,11 +14,11 @@ public class OperationDto {
     private BigDecimal unitCost;
 
     @JsonProperty("quantity")
-    private Long quantity;
+    private BigDecimal quantity;
 
     public OperationDto() {}
 
-    public OperationDto(OperationType operationType, BigDecimal unitCost, Long quantity) {
+    public OperationDto(OperationType operationType, BigDecimal unitCost, BigDecimal quantity) {
         this.operationType = operationType;
         this.unitCost = unitCost;
         this.quantity = quantity;
@@ -32,7 +32,7 @@ public class OperationDto {
         return unitCost;
     }
 
-    public Long getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 }
